+++ firmware-stm32f407-micro-ros/hyperdog_uros_interfaces/include/hyperdog_uros_interfaces/srv/detail/set_motor_cmd_frequency__struct.h
@@ -20,7 +20,7 @@ extern "C"
 // Struct defined in srv/SetMotorCmdFrequency in the package hyperdog_uros_interfaces.
 typedef struct hyperdog_uros_interfaces__srv__SetMotorCmdFrequency_Request
 {
-  double data;
+  float data;
 } hyperdog_uros_interfaces__srv__SetMotorCmdFrequency_Request;
 
 // Struct for a sequence of hyperdog_uros_interfaces__srv__SetMotorCmdFrequency_Request.
